Move 'supportsOptionalIngress' to a more reasonable location
