+++ src/components/MAPI/apps/ClusterDetailApps.tsx
@@ -14,9 +14,9 @@ import React, { useLayoutEffect, useMemo, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { AppConstants } from 'shared/constants';
 import { AppsRoutes } from 'shared/constants/routes';
+import { supportsOptionalIngress } from 'stores/cluster/utils';
 import { selectCluster } from 'stores/main/actions';
 import { getProvider } from 'stores/main/selectors';
-import { supportsOptionalIngress } from 'stores/nodepool/utils';
 import styled from 'styled-components';
 import { FlashMessageType } from 'styles';
 import useSWR from 'swr';
