+++ src/stores/cluster/utils.ts
@@ -8,7 +8,6 @@ import {
   supportsAlikeInstances,
   supportsNodePoolAutoscaling,
   supportsNodePoolSpotInstances,
-  supportsOptionalIngress,
 } from 'stores/nodepool/utils';
 import { isPreRelease } from 'stores/releases/utils';
 import { IState } from 'stores/state';
@@ -459,4 +458,19 @@ export function reconcileClustersAwaitingUpgrade(
   }
 
   return awaitingUpgrade;
+}
+
+export function supportsOptionalIngress(
+  provider: PropertiesOf<typeof Providers>,
+  releaseVersion: string
+): boolean {
+  switch (true) {
+    case provider === Providers.AWS && compare(releaseVersion, '10.1.0') >= 0:
+    case provider === Providers.AZURE && compare(releaseVersion, '12.0.0') >= 0:
+    case provider === Providers.KVM && compare(releaseVersion, '12.2.0') >= 0:
+      return true;
+
+    default:
+      return false;
+  }
 }
