+++ src/stores/nodepool/utils.ts
@@ -47,18 +47,3 @@ export function supportsAlikeInstances(
       return false;
   }
 }
-
-export function supportsOptionalIngress(
-  provider: PropertiesOf<typeof Providers>,
-  releaseVersion: string
-): boolean {
-  switch (true) {
-    case provider === Providers.AWS && compare(releaseVersion, '10.1.0') >= 0:
-    case provider === Providers.AZURE && compare(releaseVersion, '12.0.0') >= 0:
-    case provider === Providers.KVM && compare(releaseVersion, '12.2.0') >= 0:
-      return true;
-
-    default:
-      return false;
-  }
-}
